Allow Ruby 3 support
Currently, any upstream codebase using this gem is unable to keep pace with the latest Ruby development because it only supports Ruby 2.x. This commit bumps the net-http-persistent dependency to allow both 3.x (>= 3.1) and 4.x, the latter of which supports Ruby 3. That is half the battle. The other half: this commit also changes method calls to use keyword arguments instead of hashes.

I ran the test suite under both Ruby 3.0 and 2.7 and it passed under both, so this should be backwards compatible, though fair warning: I did not test under other 2.x rubies. (I am doing this work to let the team at Stanford upgrade SearchWorks to Ruby 3.0.)